companion/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
python-dotenv>=1.0.0
httpx>=0.26.0

# LLM clients (install only what you need)
google-genai>=1.0.0
anthropic>=0.39.0
# DeepSeek uses httpx (already included above)
